rename connectToStore to connectAdvanced
